requirements.txt
Flask
flask_sqlalchemy
python-dotenv
gunicorn
werkzeug
